add new options and settings

+++ console/migrations/m220123_114651_add_shop_new_function.php
@@ -12,7 +12,7 @@ class m220123_114651_add_shop_new_function extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn('{{%shop_products}}', 'weight', $this->float()->defaultValue(0)->after('rating'));
+        $this->addColumn('{{%shop_products}}', 'weight', $this->integer(11)->defaultValue(0)->after('rating'));
         $this->addColumn('{{%settings}}', 'custom_style', $this->text()->after('link_to_instagram'));
         $this->addColumn('{{%settings_lng}}', 'logo_b', $this->string(255));
         $this->addColumn('{{%settings_lng}}', 'logo_w', $this->string(255));
